Relocated arc_path from settings to common
