+++ arc/main.py
@@ -23,7 +23,7 @@ from rmgpy.species import Species
 
 import arc.rmgdb as rmgdb
 from arc.common import (VERSION,
-                        time_lapse,
+                        arc_path,
                         check_ess_settings,
                         get_logger,
                         globalize_path,
@@ -31,6 +31,7 @@ from arc.common import (VERSION,
                         initialize_log,
                         log_footer,
                         save_yaml_file,
+                        time_lapse,
                         )
 from arc.exceptions import InputError, SettingsError, SpeciesError
 from arc.level import Level
@@ -38,7 +39,7 @@ from arc.job.ssh import SSHClient
 from arc.processor import process_arc_project
 from arc.reaction import ARCReaction
 from arc.scheduler import Scheduler
-from arc.settings import arc_path, default_levels_of_theory, servers, valid_chars, default_job_types
+from arc.settings import default_levels_of_theory, servers, valid_chars, default_job_types
 from arc.species.species import ARCSpecies
 from arc.utils.scale import determine_scaling_factors
