+++ arc/mainTest.py
@@ -9,9 +9,10 @@ import os
 import shutil
 import unittest
 
+from arc.common import arc_path
 from arc.exceptions import InputError
 from arc.main import ARC, process_adaptive_levels
-from arc.settings import arc_path, servers
+from arc.settings import servers
 from arc.species.species import ARCSpecies
