+++ arc/plotterTest.py
@@ -10,7 +10,7 @@ import shutil
 import unittest
 
 import arc.plotter as plotter
-from arc.settings import arc_path
+from arc.common import arc_path
 from arc.species.converter import str_to_xyz
 from arc.species.species import ARCSpecies
