+++ arc/ts/run_autotst.py
@@ -26,9 +26,8 @@ except ImportError:
     except ImportError:
         has_auto_tst = False
 
-from arc.common import get_logger
+from arc.common import arc_path, get_logger
 from arc.exceptions import TSError
-from arc.settings import arc_path
 from arc.species.converter import xyz_from_data, xyz_to_str
