+++ arc/utils/delete.py
@@ -13,10 +13,11 @@ import argparse
 import csv
 import os
 
+from arc.common import arc_path
 from arc.exceptions import InputError
 from arc.job.local import delete_all_local_arc_jobs
 from arc.job.ssh import delete_all_arc_jobs
-from arc.settings import arc_path, servers
+from arc.settings import servers
 
 
 def parse_command_line_arguments(command_line_args=None):
