+++ ipython/Tools/demo/arc_demo_input.yml
@@ -1,0 +1,196 @@
+T_count: 50
+T_max: null
+T_min: null
+allow_nonisomorphic_2d: false
+arkane_level_of_theory:
+  basis: augccpvtz
+  compatible_ess: &id002
+  - gaussian
+  - terachem
+  method: wb97xd
+  method_type: dft
+  software: gaussian
+calc_freq_factor: true
+compute_transport: false
+conformer_level:
+  basis: def2svp
+  compatible_ess:
+  - gaussian
+  - terachem
+  method: wb97xd
+  method_type: dft
+  software: gaussian
+e_confs: 5.0
+ess_settings:
+  gaussian:
+  - local
+freq_level:
+  basis: def2tzvp
+  compatible_ess: &id001
+  - gaussian
+  - terachem
+  method: wb97xd
+  method_type: dft
+  software: gaussian
+freq_scale_factor: 0.988
+irc_level:
+  basis: def2tzvp
+  compatible_ess:
+  - gaussian
+  - terachem
+  method: wb97xd
+  method_type: dft
+  software: gaussian
+job_memory: 14
+job_types:
+  bde: false
+  conformers: true
+  fine: true
+  freq: true
+  irc: true
+  onedmin: false
+  opt: true
+  orbitals: false
+  rotors: true
+  sp: true
+kinetics_adapter: Arkane
+max_job_time: 120
+n_confs: 10
+opt_level:
+  basis: def2tzvp
+  compatible_ess: *id001
+  method: wb97xd
+  method_type: dft
+  software: gaussian
+output: {}
+project: ArcDemo
+reactions:
+- charge: 0
+  family: null
+  family_own_reverse: 0
+  index: 0
+  label: peroxyl_rad <=> ch2_rad
+  long_kinetic_description: ''
+  multiplicity: null
+  p_species: []
+  products:
+  - ch2_rad
+  r_species: []
+  reactants:
+  - peroxyl_rad
+  ts_label: null
+  ts_methods: []
+  ts_xyz_guess:
+  - ts1016.log
+running_jobs: {}
+scan_level:
+  basis: def2tzvp
+  compatible_ess: *id001
+  method: wb97xd
+  method_type: dft
+  software: gaussian
+sp_level:
+  basis: augccpvtz
+  compatible_ess: *id002
+  method: wb97xd
+  method_type: dft
+  software: gaussian
+species:
+- arkane_file: null
+  bond_corrections:
+    C-C: 2
+    C-H: 11
+    C-N: 4
+    C-O: 1
+    H-N: 2
+    O-O: 1
+  charge: 0
+  compute_thermo: true
+  consider_all_diastereomers: true
+  force_field: MMFF94s
+  is_ts: false
+  label: peroxyl_rad
+  long_thermo_description: 'Bond corrections: {''C-O'': 1, ''O-O'': 1, ''C-N'': 4,
+    ''C-H'': 11, ''C-C'': 2, ''H-N'': 2}
+
+    '
+  mol: |
+    multiplicity 2
+    1  N u0 p1 c0 {2,S} {10,S} {11,S}
+    2  C u0 p0 c0 {1,S} {3,S} {12,S} {13,S}
+    3  C u0 p0 c0 {2,S} {4,S} {14,S} {15,S}
+    4  C u0 p0 c0 {3,S} {5,S} {7,S} {16,S}
+    5  O u0 p2 c0 {4,S} {6,S}
+    6  O u1 p2 c0 {5,S}
+    7  N u0 p1 c0 {4,S} {8,S} {9,S}
+    8  C u0 p0 c0 {7,S} {17,S} {18,S} {19,S}
+    9  C u0 p0 c0 {7,S} {20,S} {21,S} {22,S}
+    10 H u0 p0 c0 {1,S}
+    11 H u0 p0 c0 {1,S}
+    12 H u0 p0 c0 {2,S}
+    13 H u0 p0 c0 {2,S}
+    14 H u0 p0 c0 {3,S}
+    15 H u0 p0 c0 {3,S}
+    16 H u0 p0 c0 {4,S}
+    17 H u0 p0 c0 {8,S}
+    18 H u0 p0 c0 {8,S}
+    19 H u0 p0 c0 {8,S}
+    20 H u0 p0 c0 {9,S}
+    21 H u0 p0 c0 {9,S}
+    22 H u0 p0 c0 {9,S}
+  multiplicity: 2
+  number_of_rotors: 0
+- arkane_file: null
+  bond_corrections:
+    C-C: 2
+    C-H: 10
+    C-N: 4
+    C-O: 1
+    H-N: 2
+    H-O: 1
+    O-O: 1
+  charge: 0
+  compute_thermo: true
+  consider_all_diastereomers: true
+  force_field: MMFF94s
+  is_ts: false
+  label: ch2_rad
+  long_thermo_description: 'Bond corrections: {''C-O'': 1, ''H-O'': 1, ''O-O'': 1,
+    ''C-N'': 4, ''C-H'': 10, ''C-C'': 2, ''H-N'': 2}
+
+    '
+  mol: |
+    multiplicity 2
+    1  N u0 p1 c0 {2,S} {12,S} {13,S}
+    2  C u0 p0 c0 {1,S} {3,S} {14,S} {15,S}
+    3  C u0 p0 c0 {2,S} {4,S} {16,S} {17,S}
+    4  C u0 p0 c0 {3,S} {5,S} {7,S} {18,S}
+    5  O u0 p2 c0 {4,S} {6,S}
+    6  O u0 p2 c0 {5,S} {19,S}
+    7  N u0 p1 c0 {4,S} {8,S} {9,S}
+    8  C u0 p0 c0 {7,S} {20,S} {21,S} {22,S}
+    9  C u1 p0 c0 {7,S} {10,S} {11,S}
+    10 H u0 p0 c0 {9,S}
+    11 H u0 p0 c0 {9,S}
+    12 H u0 p0 c0 {1,S}
+    13 H u0 p0 c0 {1,S}
+    14 H u0 p0 c0 {2,S}
+    15 H u0 p0 c0 {2,S}
+    16 H u0 p0 c0 {3,S}
+    17 H u0 p0 c0 {3,S}
+    18 H u0 p0 c0 {4,S}
+    19 H u0 p0 c0 {6,S}
+    20 H u0 p0 c0 {8,S}
+    21 H u0 p0 c0 {8,S}
+    22 H u0 p0 c0 {8,S}
+  multiplicity: 2
+  number_of_rotors: 0
+thermo_adapter: Arkane
+ts_guess_level:
+  basis: def2svp
+  compatible_ess:
+  - gaussian
+  - terachem
+  method: wb97xd
+  method_type: dft
+  software: gaussian
